requirements.txt
expertai_nlapi==2.4.1
notion_client==1.0.0
python-dotenv==0.21.0
tqdm==4.64.1
